build.py changes.

+++ build.py
@@ -10,9 +10,15 @@ if __name__ == "__main__":
     if not os.path.exists("lib"):
         os.mkdir("lib")
     for module in MODULES:
+        print(f"Building nstd_{module}...")
         os.chdir(f"src/std/{module}")
-        os.system("cargo build --release")
+        os.system("cargo build --release --quiet")
         if sys.platform == "win32":
-            os.rename(
-                f"target/release/nstd_{module}.lib", f"../../../lib/nstd_{module}.lib")
+            if os.path.exists(f"../../../lib/nstd_{module}.lib"):
+                os.replace(
+                    f"target/release/nstd_{module}.lib", f"../../../lib/nstd_{module}.lib")
+            else:
+                os.rename(
+                    f"target/release/nstd_{module}.lib", f"../../../lib/nstd_{module}.lib")
         os.chdir("../../../")
+        print(f"Finished nstd_{module}.")
